+++ src/eulertools/subcommands/generate.py
@@ -10,6 +10,8 @@ from eulertools.lib.utils import (
 
 
 class Generate:
+    __slots__ = ("languages", "problems")
+
     def __init__(self, languages: list[Language], problems: list[Problem]):
         self.languages = languages
         self.problems = problems