+++ src/eulertools/subcommands/statement.py
@@ -4,6 +4,8 @@ from eulertools.lib.utils import Problem, get_statement
 
 
 class Statement:
+    __slots__ = ("problems", "show_hints")
+
     def __init__(self, problems: list[Problem], *, show_hints: bool):
         self.problems = problems
         self.show_hints = show_hints
@@ -11,6 +13,7 @@ class Statement:
     def run(self) -> None:
         for problem in self.problems:
             self.show_statement(problem)
+            print()
 
     def show_statement(self, problem: Problem) -> None:
         statement = get_statement(problem)["common"]